Updated to use Custom Blossom * Added custom blossom injector dependency * Updated dependencies * Moved github update checker (adding more update checkers) * Updated .idea files * Removed simplecoreapi properties to use injected variables
