+++ build.gradle
@@ -2,7 +2,7 @@ plugins {
     id 'org.jetbrains.kotlin.jvm' version '1.7.0'
     id 'maven-publish'
     id 'com.github.johnrengelman.shadow' version '7.1.2'
-    id 'net.kyori.blossom' version '1.2.0'
+    id 'cl.franciscosolis.blossom-extended' version '1.3.1'
     id 'org.jetbrains.dokka' version '1.6.21'
 }
 
@@ -65,8 +65,6 @@ java {
 }
 
 blossom {
-    replaceTokenIn('src/main/kotlin/xyz/theprogramsrc/simplecoreapi/velocity/VelocityLoader.kt')
-
     replaceToken '@name@', rootProject.name
     replaceToken '@version@', project.version.toString()
     replaceToken '@description@', project.description
