+++ src/main/kotlin/xyz/theprogramsrc/simplecoreapi/global/SimpleCoreAPI.kt
@@ -3,6 +3,7 @@ package xyz.theprogramsrc.simplecoreapi.global
 import xyz.theprogramsrc.simplecoreapi.global.module.ModuleManager
 import xyz.theprogramsrc.simplecoreapi.global.utils.ILogger
 import xyz.theprogramsrc.simplecoreapi.global.utils.SoftwareType
+import xyz.theprogramsrc.simplecoreapi.global.utils.update.GitHubUpdateChecker
 import java.util.*
 
 /**
@@ -27,12 +28,6 @@ class SimpleCoreAPI(logger: ILogger) {
     val moduleManager: ModuleManager?
 
     /**
-     * SimpleCoreAPI Properties
-     * @return The [Properties] of SimpleCoreAPI
-     */
-    val props: Properties = Properties()
-
-    /**
      * The [SoftwareType] type running on the server
      * @return The [SoftwareType] the server is running
      */
@@ -40,15 +35,12 @@ class SimpleCoreAPI(logger: ILogger) {
 
     init {
         instance = this
-        val resource = SimpleCoreAPI::class.java.getResource("/simplecoreapi.properties")
-        if (resource != null) {
-            props.load(resource.openStream())
-        }
 
         logger.info("SimpleCoreAPI v${getVersion()} - Git Commit: ${getShortHash()}")
         if (getFullHash() != "unknown") {
             GitHubUpdateChecker(logger, "TheProgramSrc/SimpleCoreAPI", getVersion()).checkWithPrint()
         }
+
         softwareType = SoftwareType.values().firstOrNull { it.check() } ?: SoftwareType.UNKNOWN
         if(softwareType != SoftwareType.UNKNOWN && softwareType.display != null) {
             logger.info("Running API with software ${softwareType.display}")
@@ -60,19 +52,19 @@ class SimpleCoreAPI(logger: ILogger) {
      * Gets the short version of the commit hash
      * @return The short commit hash
      */
-    fun getShortHash(): String = props.getProperty("git-short", "unknown")
+    fun getShortHash(): String = "@git_short@"
 
     /**
      * Gets the full version of the commit hash
      * @return The full commit hash
      */
-    fun getFullHash(): String = props.getProperty("git-full", "unknown")
+    fun getFullHash(): String = "@git_full@"
 
     /**
      * Gets the version of SimpleCoreAPI
      * @return The version of SimpleCoreAPI
      */
-    fun getVersion(): String = props.getProperty("version", "unknown")
+    fun getVersion(): String = "@version@"
 
     /**
      * Checks if the current [SoftwareType] is the one specified
