--- src/main/resources/simplecoreapi.properties
@@ -1,3 +1,0 @@
-git-full=@git_full@
-git-short=@git_short@
-version=@version@
